+++ tests/models/identifiers/test_identifiers.py
@@ -1,0 +1,38 @@
+from lxml import etree
+
+from caselawclient.models.identifiers import Identifier, IdentifierSchema
+
+
+class TestIdentifierSchema(IdentifierSchema):
+    __test__ = False
+
+    name = "Test Schema"
+    namespace = "test"
+
+
+class TestIdentifier(Identifier):
+    __test__ = False
+
+    schema = TestIdentifierSchema
+
+
+class TestIdentifierBase:
+    def test_uuid_setting(self):
+        """Ensure that if a UUID is provided when initialising an Identifier that it is properly stored."""
+        identifier = Identifier(uuid="2d80bf1d-e3ea-452f-965c-041f4399f2dd", value="TEST-123")
+        assert identifier.uuid == "2d80bf1d-e3ea-452f-965c-041f4399f2dd"
+
+    def test_xml_representation(self):
+        """Ensure that an identifer generates the expected representation of itself as XML."""
+        identifier = TestIdentifier(uuid="2d80bf1d-e3ea-452f-965c-041f4399f2dd", value="TEST-123")
+
+        expected_xml = """
+            <identifier>
+                <namespace>test</namespace>
+                <uuid>2d80bf1d-e3ea-452f-965c-041f4399f2dd</uuid>
+                <value>TEST-123</value>
+            </identifier>
+        """
+        assert etree.canonicalize(identifier.as_xml_tree, strip_text=True) == etree.canonicalize(
+            etree.fromstring(expected_xml), strip_text=True
+        )